Modify Code Structure

+++ app/helper.php
@@ -1,4 +1,7 @@
 <?php
+/////////////////////////////////
+/////////HELPER METHODS//////////
+/////////////////////////////////
 
 use Illuminate\Support\Facades\Storage;
 // use hisorange\BrowserDetect\Facade as BrowserDetect;
@@ -12,7 +15,7 @@ function apiResponse($success, $message, $data, $statusCode = 200)
     ], $statusCode);
 }
 
-function removeCustomTags($content, $tags = ['script','iframe','style'])
+function removeCustomTags($content, $tags = ['script', 'iframe', 'style'])
 {
     foreach ($tags as $tag) {
         $content = str_replace('<' . $tag, '&lt;' . $tag, $content);
@@ -32,7 +35,7 @@ function uploadFile($request, $uploadedFileName, $newFileName, $oldPath = "", $f
     }
     $savePath = $path . '/' . ($newFileName ? $newFileName . '.' . $file->getClientOriginalExtension() : $file->getClientOriginalName());
     $UploadedFile = Storage::disk('public')->put($savePath, file_get_contents($file));
-    if ($UploadedFile && $oldPath) {
+    if ($UploadedFile && $oldPath && $savePath != $oldPath) {
         if (Storage::disk('public')->exists($oldPath)) {
             Storage::disk('public')->delete($oldPath);
         }
@@ -128,3 +131,7 @@ function getIp()
 {
     return request()->ip();
 }
+
+/////////////////////////////////
+///////END HELPER METHODS////////
+/////////////////////////////////
